backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlmodel>=0.0.19
pydantic-settings>=2.1.0
celery[redis]>=5.3.6
redis>=5.0.1
httpx>=0.25.2
websockets>=12.0
openai>=1.3.0
yfinance>=0.2.33
pandas>=2.1.0
ta>=0.10.2
python-multipart>=0.0.6
